Make bin a minimal wrapper
